requirements/base.txt
asgiref==3.6.0
build==0.10.0
cfgv==3.3.1
click==8.1.3
distlib==0.3.6
Django==4.2
django-cors-headers==3.10.0
django-extensions==3.2.1
django-filter==23.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-spectacular==0.26.2
factory-boy==3.2.1
filelock==3.12.0
flake8==6.0.0
gunicorn==21.2.0
identify==2.5.22
isort==5.12.0
mccabe==0.7.0
mypy-extensions==1.0.0
nodeenv==1.7.0
packaging==23.1
pathspec==0.11.1
Pillow==9.5.0
pip-tools==6.13.0
platformdirs==3.2.0
pre-commit==3.2.2
pypdf==3.9.1
pycodestyle==2.10.0
pyflakes==3.0.1
pyproject_hooks==1.0.0
python-barcode==0.14.0
python-dotenv==1.0.0
pytz==2023.3
PyYAML==6.0
sqlparse==0.4.4
virtualenv==20.21.0
whitenoise==6.2.0
django-storages==1.14.2
boto3==1.28.80
request==2.31.0